feat: add premium handling and proper test instance handling

+++ events/guildDelete.js
@@ -1,4 +1,7 @@
 module.exports = async (guild, bot) => {
+	if(process.env.TESTING && guild.id !== process.env.TEST_GUILD)
+		return;
+
 	try {
 		await bot.stores.serverRoles.deleteAll(guild.id);
 	} catch(e) {
